jd01_02 refactor architecture
