+++ src/by/it/shcharbunou/jd02_01/multithreaded_store/services/store_services/GoodsProducer.java
@@ -1,0 +1,6 @@
+package by.it.shcharbunou.jd02_01.multithreaded_store.services.store_services;
+
+public class GoodsProducer {
+
+
+}
