+++ src/by/it/shcharbunou/jd02_01/multithreaded_store/services/store_services/Store.java
@@ -1,4 +1,4 @@
-package by.it.shcharbunou.jd02_01.multithreaded_store.services;
+package by.it.shcharbunou.jd02_01.multithreaded_store.services.store_services;
 
 public class Store implements Runnable {
